libefs: Change to C99 and improve formatting

+++ libefs/source/aes.c
@@ -262,7 +262,7 @@ static void ShiftRows(
 static uint8_t xtime(
     uint8_t x
 ) {
-    return ((x << 1) ^ (((x >> 7) & 1) * 0x1b));
+    return (uint8_t) ((x << 1) ^ (((x >> 7) & 1) * 0x1b));
 }
 
 // MixColumns function mixes the columns of the state matrix
